[Improvement][api] Introduce users service interface for clear code
